+++ tests/TODO/functional/test_pages.py
@@ -5,71 +5,6 @@ import pytest
 from webapp.app import User, app, db
 
 
-def test_generate_mealplan(test_client):
-    """
-    GIVEN a Flask application configured for testing
-    WHEN the '/mealplan' page is posted to (POST) when the user enters health requirements data
-
-    """
-    response = test_client.post(
-        "/mealplan",
-        data=dict(
-            Calories="129.7",
-            Carbs="57.6",
-            Proteins="68.5",
-            fiber="75.9",
-            caloriesbreakfastamount=".7",
-            calorieslunchamount=".2",
-            carbsbreakfastamount=".3",
-            carbslunchamount=".6",
-            proteinsbreakfastamount=".5",
-            proteinslunchamount=".1",
-        ),
-        follow_redirects=True,
-    )
-    assert b"Personal Meal Plan Recommendations" in response.data
-    assert response.status_code == 200
-
-
-@pytest.mark.xfail(reason="not yet implemented")
-def test_generate_exercise_plan(test_client):
-    """
-    GIVEN a Flask application configured for testing
-    WHEN the '/mealplan' page is posted to (POST) when the user enters health requirements data
-
-    """
-    response = test_client.post(
-        "/exerciseplan",
-        data=dict(sunday=True, friday=True,
-                  intensity="8", glutes=True, chest=True),
-        follow_redirects=True,
-    )
-    assert b"Personal Exercise Plan Recommendations" in response.data
-    assert response.status_code == 200
-
-
-@pytest.mark.xfail(reason="not yet implemented")
-def test_add_exercise(test_client):
-    """
-    GIVEN a Flask application configured for testing
-    WHEN the '/mealplan' page is posted to (POST) when the user enters health requirements data
-
-    """
-    response = test_client.post(
-        "/addexercise",
-        data=dict(
-            name="newexercise",
-            sets="9",
-            reps="10",
-            intensity="7",
-            muscles=["quads", "abs", "hamstrings"],
-        ),
-        follow_redirects=True,
-    )
-    assert b"newexercise" in response.data
-    assert response.status_code == 200
-
-
 @pytest.mark.xfail(reason="not yet implemented")
 def test_add_food_item(test_client):
     """
