+++ tests/UC_3_plan_meals_for_week/End_to_end_tests/test_UC3_E2E_1.py
@@ -63,6 +63,7 @@ def test_UC3_E2E_1(test_client, login_default_user):
     """
     scaling_factor = 0.2
 
+    # [-----|-----|----]
     # Send user constraints
     data = dict(
         Calories="129.7",
@@ -70,11 +71,11 @@ def test_UC3_E2E_1(test_client, login_default_user):
         Proteins="68.5",
         fiber="75.9",
         caloriesbreakfastamount=".7",
-        calorieslunchamount=".2",
+        calorieslunchamount=".9",
         carbsbreakfastamount=".3",
-        carbslunchamount=".6",
+        carbslunchamount=".9",
         proteinsbreakfastamount=".5",
-        proteinslunchamount=".1",
+        proteinslunchamount=".6",
     )
 
     response = test_client.post(
@@ -93,7 +94,7 @@ def test_UC3_E2E_1(test_client, login_default_user):
     tbl_items = [
         x for x in [x for x in table[0].children if x != "\n"][0] if x != "\n"
     ][1:]
-    partitioned = [tbl_items[4 * i: 4 * (i + 1)] for i in range(3)]
+    partitioned = [tbl_items[4 * i : 4 * (i + 1)] for i in range(3)]
     dfs = []
     # replaces last index of partitioned with series
     for A in partitioned:
@@ -117,11 +118,9 @@ def test_UC3_E2E_1(test_client, login_default_user):
         "",
         pd.Series(
             [
-                float(data["Calories"]) *
-                float(data["caloriesbreakfastamount"]),
+                float(data["Calories"]) * float(data["caloriesbreakfastamount"]),
                 float(data["Carbs"]) * float(data["carbsbreakfastamount"]),
-                float(data["Proteins"]) *
-                float(data["proteinsbreakfastamount"]),
+                float(data["Proteins"]) * float(data["proteinsbreakfastamount"]),
             ],
             index=["Calories", "Carbs", "Proteins"],
         ),
@@ -140,17 +139,9 @@ def test_UC3_E2E_1(test_client, login_default_user):
         ),
     )
 
-    derived_calories_dinner: float = 1 - (
-        float(data["calorieslunchamount"]) +
-        float(data["caloriesbreakfastamount"])
-    )
-    derived_carbs_dinner: float = 1 - (
-        float(data["carbslunchamount"]) + float(data["carbsbreakfastamount"])
-    )
-    derived_proteins_dinner: float = 1 - (
-        float(data["proteinslunchamount"]) +
-        float(data["proteinsbreakfastamount"])
-    )
+    derived_calories_dinner: float = 1 - (float(data["calorieslunchamount"]))
+    derived_carbs_dinner: float = 1 - (float(data["carbslunchamount"]))
+    derived_proteins_dinner: float = 1 - (float(data["proteinslunchamount"]))
     assert derived_proteins_dinner <= 1 and derived_proteins_dinner >= 0
     assert derived_carbs_dinner <= 1 and derived_carbs_dinner >= 0
     assert derived_proteins_dinner <= 1 and derived_proteins_dinner >= 0
@@ -170,19 +161,18 @@ def test_UC3_E2E_1(test_client, login_default_user):
 
     # Breakfast
     Breakfast = Meal(
-        dfs[0][0], dfs[0][1], dfs[0][2], dfs[0][3][[
-            "Calories", "Carbs", "Protein"]]
+        dfs[0][0], dfs[0][1], dfs[0][2], dfs[0][3][["Calories", "Carbs", "Protein"]]
     )
     # Lunch
     Lunch = Meal(
-        dfs[1][0], dfs[1][1], dfs[1][2], dfs[1][3][[
-            "Calories", "Carbs", "Protein"]]
+        dfs[1][0], dfs[1][1], dfs[1][2], dfs[1][3][["Calories", "Carbs", "Protein"]]
     )
     # Dinner
     Dinner = Meal(
-        dfs[2][0], dfs[2][1], dfs[2][2], dfs[2][3][[
-            "Calories", "Carbs", "Protein"]]
+        dfs[2][0], dfs[2][1], dfs[2][2], dfs[2][3][["Calories", "Carbs", "Protein"]]
     )
+
+    breakpoint()
 
     def _generic_meal_check(Meal, Expected_Meal, scaling_factor):
         # calories
